ref #4157 

Think this should implement the approach from: https://lmsys.org/blog/2023-11-21-lookahead-decoding/

- The approach requires large batches to be decoded, which in turn requires a lot of FLOPS even for single stream
- Haven't done extensive tests on NVIDIA GPUs yet, but on Apple Silicon I don't observe speed-ups with 7B F16 models, likely due to small compute
- As usual quantum models would require efficient batch CUDA kernels which we haven't implemented yet, so I don't expect speed-up there as well for the moment

In `llama.cpp` the required attention mask is achieved with the following `llama_batch`:

```
        // Example for W = 5, N = 4, G = 2:
        // (I = input, L = lookahead, V = verification)
        //
        // Batch:  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20
        // T:        -2 -2 -2 -2 -1 -1 -1 -1 -1  0  0  0  0  0  0
        // Info:   I  L  L  L  L  L  L  L  L  L  L  L  L  L  L  V  V  V  V  V  V
        // Pos:    0  1  2  3  4  1  2  3  4  5  2  3  4  5  6  1  2  3  1  2  3   (+ n_past)
        // Logits: 1  0  0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1  1  1  1
        // ---------------------------------------------------------------------
        // Seq:    0
        //         1              1              1
        //         2  2              2              2
        //         3  3  3              3              3
        //         4  4  4  4              4              4
        //         5  5  5  5  5              5              5
        //         6                                            6  6  6
        //         7                                                     7  7  7
        // ---------------------------------------------------------------------
        //                                       |  |  |  |  |  |  |  |  |  |  |
        //                                       V  V  V  V  V  |  |  |  |  |  |
        //                                         j_tokens     |  |  |  |  |  |
        //                                                      V  V  V  V  V  V
        //                                                             id
```

TODO:

- [ ] Pass `W`, `N`, `G` from command-line (currently hardcoded)
- [ ] Decode the verification n-grams via tree-based decoding. Since most of the n-grams share common prefixes, we can ramp-up the number `G` at no additional computational cost. See `speculative` example for a demonstration of tree-based decoding

```bash
make -j && ./bin/lookahead -m ../models/codellama-7b/ggml-model-f16.gguf -p "// network server implemented in C\n// author: Peter Hacker\n\n#include" -e -ngl 99 -t 4 -n 512 -c 4096 --temp 0.0
```

<img width="911" alt="image" src="https://github.com/ggerganov/llama.cpp/assets/1991296/297d2442-c253-4ff5-9756-3419a57636e0">