@slaren 2023-11-26
I don't see a speedup with CUDA/3090 Ti, but it is not too far off, about 10% slower than `main`. Maybe after tuning the parameters it could be an speedup.