@Galunid 2023-11-26
On average `lookahead` seems to perform better than `main` on A100

```
### LOOKAHEAD: The best music is ###
llama_print_timings:        load time =    4026.47 ms
llama_print_timings:      sample time =    2109.07 ms /  5929 runs   (    0.36 ms per token,  2811.19 tokens per second)
llama_print_timings: prompt eval time =   23935.60 ms / 29692 tokens (    0.81 ms per token,  1240.50 tokens per second)
llama_print_timings:        eval time =      33.45 ms /     1 runs   (   33.45 ms per token,    29.90 tokens per second)
llama_print_timings:       total time =   29742.02 ms
### main: The best music is ###
llama_print_timings:        load time =    2744.81 ms
llama_print_timings:      sample time =     189.53 ms /   512 runs   (    0.37 ms per token,  2701.45 tokens per second)
llama_print_timings: prompt eval time =      26.76 ms /     5 tokens (    5.35 ms per token,   186.87 tokens per second)
llama_print_timings:        eval time =    8260.66 ms /   511 runs   (   16.17 ms per token,    61.86 tokens per second)
llama_print_timings:       total time =   10722.21 ms
### LOOKAHEAD: Here's a short summary of Poland's history ###
llama_print_timings:        load time =    3093.91 ms
llama_print_timings:      sample time =    2080.90 ms /  5854 runs   (    0.36 ms per token,  2813.20 tokens per second)
llama_print_timings: prompt eval time =   23482.37 ms / 30067 tokens (    0.78 ms per token,  1280.41 tokens per second)
llama_print_timings:        eval time =      11.79 ms /     1 runs   (   11.79 ms per token,    84.84 tokens per second)
llama_print_timings:       total time =   29430.38 ms
### main: Here's a short summary of Poland's history ###
llama_print_timings:        load time =    5244.26 ms
llama_print_timings:      sample time =     307.42 ms /   512 runs   (    0.60 ms per token,  1665.50 tokens per second)
llama_print_timings: prompt eval time =      35.39 ms /    12 tokens (    2.95 ms per token,   339.06 tokens per second)
llama_print_timings:        eval time =   10204.68 ms /   511 runs   (   19.97 ms per token,    50.08 tokens per second)
llama_print_timings:       total time =   12466.79 ms
### LOOKAHEAD: The main points of the constitution of the United States are: ###
llama_print_timings:        load time =    2356.03 ms
llama_print_timings:      sample time =    1463.57 ms /  5194 runs   (    0.28 ms per token,  3548.86 tokens per second)
llama_print_timings: prompt eval time =   20358.28 ms / 27476 tokens (    0.74 ms per token,  1349.62 tokens per second)
llama_print_timings:        eval time =      11.76 ms /     1 runs   (   11.76 ms per token,    85.06 tokens per second)
llama_print_timings:       total time =   25264.51 ms
### main: The main points of the constitution of the United States are: ###
llama_print_timings:        load time =    2544.58 ms
llama_print_timings:      sample time =     166.50 ms /   512 runs   (    0.33 ms per token,  3075.08 tokens per second)
llama_print_timings: prompt eval time =      34.13 ms /    13 tokens (    2.63 ms per token,   380.87 tokens per second)
llama_print_timings:        eval time =    9196.50 ms /   511 runs   (   18.00 ms per token,    55.56 tokens per second)
llama_print_timings:       total time =   10841.33 ms
### LOOKAHEAD: Here's a tic-tac-toe in python: ###
llama_print_timings:        load time =    2247.46 ms
llama_print_timings:      sample time =    1400.48 ms /  4969 runs   (    0.28 ms per token,  3548.08 tokens per second)
llama_print_timings: prompt eval time =   20073.95 ms / 27371 tokens (    0.73 ms per token,  1363.51 tokens per second)
llama_print_timings:        eval time =      11.77 ms /     1 runs   (   11.77 ms per token,    84.97 tokens per second)
llama_print_timings:       total time =   24821.21 ms
### main: Here's a tic-tac-toe in python: ###
llama_print_timings:        load time =    2389.40 ms
llama_print_timings:      sample time =     166.72 ms /   512 runs   (    0.33 ms per token,  3071.00 tokens per second)
llama_print_timings: prompt eval time =      34.58 ms /    16 tokens (    2.16 ms per token,   462.71 tokens per second)
llama_print_timings:        eval time =    7831.22 ms /   511 runs   (   15.33 ms per token,    65.25 tokens per second)
llama_print_timings:       total time =    9455.96 ms
### LOOKAHEAD: This is the best day of my life ###
llama_print_timings:        load time =    3528.09 ms
llama_print_timings:      sample time =    1227.67 ms /  3356 runs   (    0.37 ms per token,  2733.64 tokens per second)
llama_print_timings: prompt eval time =   12738.32 ms / 16784 tokens (    0.76 ms per token,  1317.60 tokens per second)
llama_print_timings:        eval time =      17.77 ms /     1 runs   (   17.77 ms per token,    56.27 tokens per second)
llama_print_timings:       total time =   16012.46 ms
### main: This is the best day of my life ###
llama_print_timings:        load time =    2589.66 ms
llama_print_timings:      sample time =     135.30 ms /   363 runs   (    0.37 ms per token,  2682.91 tokens per second)
llama_print_timings: prompt eval time =      41.00 ms /     9 tokens (    4.56 ms per token,   219.53 tokens per second)
llama_print_timings:        eval time =    6353.26 ms /   362 runs   (   17.55 ms per token,    56.98 tokens per second)
llama_print_timings:       total time =    7766.58 ms
```

Here's the whole output: [output.txt](https://github.com/ggerganov/llama.cpp/files/13468053/output.txt)

I used simple script to run `./prompts.sh 2>&1 | tee output.txt` on base mistral model (`Q8_0`)
#### prompts.sh:
```bash
prompts=(
    "The best music is"
    "Here's a short summary of Poland's history"
    "The main points of the constitution of the United States are:"
    "Here's a tic-tac-toe in python:"
    "This is the best day of my life"
)

for prompt in "${prompts[@]}"; do
    echo "### LOOKAHEAD: $prompt ###"
    ./lookahead -m ../mistral-7b-v0.1.Q8_0.gguf -p "$prompt" -ngl 35 -n 512 -c 4096
    echo "### main: $prompt ###"
    ./main -m ../mistral-7b-v0.1.Q8_0.gguf -p "$prompt" -ngl 35 -n 512 -c 4096
done
```

If there's anything else you want tested do let me know.